myDash: rework running dots.
Programmatically draw the running dots instead of using background svgs.
This has the advantage of allowing to define the dot style from the css
style. The background-color, border-color and border-width of the
default "dot", which is typically defined by the class
'app-well-app-running-dot', is used: this makes the dots integrated with
the default color-scheme even if custom themes are used.