window-preview: backporting from 3.22
Add windowPreview class.

appIcons: add window previews to the apps menu.

The list of windows is replaced with a submenu with windows previews.

windowPreview: animate out clone when windows are closed.

windowPreview: add close button.

appIcons: fix bug with popupMenu.

If the parent actor of the popupmenu is destroyed while the menu is
open, the grab is not properly released as the function try to access
the actor.

This is an upstream bug:
https://bugzilla.gnome.org/show_bug.cgi?id=757556.

This can be safely removed unce it gets fixed upstream.

appIcons: update popupmenu on app windows change.

Add an update method to update the menu content when the application
windows change. This is desiderable as the menu is used to change the
open windows by closing the. Moreover, also repopulate the windows
preview when new windows are added.

windowsPreview: wait for a proper texture when created.

It turned out that the windows added signal is emitted before the
windows actor is acutally added. Similarly to what done in the overview
workspace, wait for a valid texture. Also note taht sometimes an empty
texture is retrieved, so as a workaround also check for this property.

appIcons: improve transition from populated to empty windows previews.

Set the menu as insensitive instead of hiding it when the last windows
of the application is destroyed. This result in a smoother transition,
without the menu being resized for example.

Add option for showing windows previews.

appIcons: use option for showing windows previews.

appIcons: fix compatibility with workspace isolation.

Make use of getInterestingWindows() instead of using all windows.  Make
use of closeAllWIndows() instead of a deprecated function.